Refactor imports to use updated libs path
