+++ libs/OpenAiCompletaions.ts
