+++ pages/api/openai.ts
@@ -1,4 +1,4 @@
-import { OpenAIStream, OpenAIStreamPayload } from "../../lib/OpenAIStream";
+import { OpenAIStream, OpenAIStreamPayload } from "../../libs/OpenAIStream";
 
 if (!process.env.NEXT_PUBLIC_OPENAI_API_KEY) {
   throw new Error("Missing env var from OpenAI");
